+++ projects/VoxelRenderer/content/PathTrace/PrimaryRay.compute.glsl
@@ -82,12 +82,12 @@ void setRayPosition(ivec3 coord, vec3 value)
 
 layout(std430, binding = 2) buffer RayDirection
 {
-    float rayDirection[];
+    float16_t rayDirection[];
 };
 
 layout(std430, binding = 3) buffer RayDirectionOut
 {
-    float rayDirectionOut[];
+    float16_t rayDirectionOut[];
 };
 
 vec3 getRayDirection(ivec3 coord)
@@ -100,9 +100,9 @@ vec3 getRayDirection(ivec3 coord)
 void setRayDirection(ivec3 coord, vec3 value)
 {
     int index = 3 * (coord.x + resolution.x * coord.y); // Stride of 3, axis order is x y z
-    rayDirectionOut[0 + index] = value.x;
-    rayDirectionOut[1 + index] = value.y;
-    rayDirectionOut[2 + index] = value.z;
+    rayDirectionOut[0 + index] = float16_t(value.x);
+    rayDirectionOut[1 + index] = float16_t(value.y);
+    rayDirectionOut[2 + index] = float16_t(value.z);
 }
 
 layout(std430, binding = 4) buffer OccupancyMap
@@ -225,16 +225,16 @@ void setFirstHitMaterial(ivec3 coord, int value)
 
 layout(std430, binding = 12) buffer SecondaryDirection
 {
-    writeonly float secondaryDirection[];
+    writeonly float16_t secondaryDirection[];
 };
 
 void setSecondaryDirection(ivec3 coord, vec4 value)
 {
     int index = 4 * (coord.x + resolution.x * (coord.y)); // Stride of 1, axis order is x y
-    secondaryDirection[index + 0] = value.x;
-    secondaryDirection[index + 1] = value.y;
-    secondaryDirection[index + 2] = value.z;
-    secondaryDirection[index + 3] = value.w;
+    secondaryDirection[index + 0] = float16_t(value.x);
+    secondaryDirection[index + 1] = float16_t(value.y);
+    secondaryDirection[index + 2] = float16_t(value.z);
+    secondaryDirection[index + 3] = float16_t(value.w);
 }
 
 uniform bool whichAccumulatingVector;
